@@ -1,208 +1,18 @@
-# Assisted by watsonx Code Assistant
-
+import warnings
+from io import BytesIO
 from pathlib import Path
 
 import pytest
-from docling_core.types.doc import DoclingDocument
-from pydantic import ValidationError
+from docling_core.types.doc import DoclingDocument, GroupItem, TextItem
 
-from docling.backend.webvtt_backend import (
-    _WebVTTCueItalicSpan,
-    _WebVTTCueTextSpan,
-    _WebVTTCueTimings,
-    _WebVTTCueVoiceSpan,
-    _WebVTTFile,
-    _WebVTTTimestamp,
-)
-from docling.datamodel.base_models import InputFormat
-from docling.datamodel.document import ConversionResult
+from docling.datamodel.base_models import DocumentStream, InputFormat
+from docling.datamodel.document import ConversionResult, _DocumentConversionInput
 from docling.document_converter import DocumentConverter
 
 from .test_data_gen_flag import GEN_TEST_DATA
 from .verify_utils import verify_document, verify_export
 
 GENERATE = GEN_TEST_DATA
-
-
-def test_vtt_cue_commponents():
-    """Test WebVTT components."""
-    valid_timestamps = [
-        "00:01:02.345",
-        "12:34:56.789",
-        "02:34.567",
-        "00:00:00.000",
-    ]
-    valid_total_seconds = [
-        1 * 60 + 2.345,
-        12 * 3600 + 34 * 60 + 56.789,
-        2 * 60 + 34.567,
-        0.0,
-    ]
-    for idx, ts in enumerate(valid_timestamps):
-        model = _WebVTTTimestamp(raw=ts)
-        assert model.seconds == valid_total_seconds[idx]
-
-    """Test invalid WebVTT timestamps."""
-    invalid_timestamps = [
-        "00:60:02.345",  # minutes > 59
-        "00:01:60.345",  # seconds > 59
-        "00:01:02.1000",  # milliseconds > 999
-        "01:02:03",  # missing milliseconds
-        "01:02",  # missing milliseconds
-        ":01:02.345",  # extra : for missing hours
-        "abc:01:02.345",  # invalid format
-    ]
-    for ts in invalid_timestamps:
-        with pytest.raises(ValidationError):
-            _WebVTTTimestamp(raw=ts)
-
-    """Test the timestamp __str__ method."""
-    model = _WebVTTTimestamp(raw="00:01:02.345")
-    assert str(model) == "00:01:02.345"
-
-    """Test valid cue timings."""
-    start = _WebVTTTimestamp(raw="00:10.005")
-    end = _WebVTTTimestamp(raw="00:14.007")
-    cue_timings = _WebVTTCueTimings(start=start, end=end)
-    assert cue_timings.start == start
-    assert cue_timings.end == end
-    assert str(cue_timings) == "00:10.005 --> 00:14.007"
-
-    """Test invalid cue timings with end timestamp before start."""
-    start = _WebVTTTimestamp(raw="00:10.700")
-    end = _WebVTTTimestamp(raw="00:10.500")
-    with pytest.raises(ValidationError) as excinfo:
-        _WebVTTCueTimings(start=start, end=end)
-    assert "End timestamp must be greater than start timestamp" in str(excinfo.value)
-
-    """Test invalid cue timings with missing end."""
-    start = _WebVTTTimestamp(raw="00:10.500")
-    with pytest.raises(ValidationError) as excinfo:
-        _WebVTTCueTimings(start=start)
-    assert "Field required" in str(excinfo.value)
-
-    """Test invalid cue timings with missing start."""
-    end = _WebVTTTimestamp(raw="00:10.500")
-    with pytest.raises(ValidationError) as excinfo:
-        _WebVTTCueTimings(end=end)
-    assert "Field required" in str(excinfo.value)
-
-    """Test with valid text."""
-    valid_text = "This is a valid cue text span."
-    span = _WebVTTCueTextSpan(text=valid_text)
-    assert span.text == valid_text
-    assert str(span) == valid_text
-
-    """Test with text containing newline characters."""
-    invalid_text = "This cue text span\ncontains a newline."
-    with pytest.raises(ValidationError):
-        _WebVTTCueTextSpan(text=invalid_text)
-
-    """Test with text containing ampersand."""
-    invalid_text = "This cue text span contains &."
-    with pytest.raises(ValidationError):
-        _WebVTTCueTextSpan(text=invalid_text)
-
-    """Test with text containing less-than sign."""
-    invalid_text = "This cue text span contains <."
-    with pytest.raises(ValidationError):
-        _WebVTTCueTextSpan(text=invalid_text)
-
-    """Test with empty text."""
-    with pytest.raises(ValidationError):
-        _WebVTTCueTextSpan(text="")
-
-    """Test that annotation validation works correctly."""
-    valid_annotation = "valid-annotation"
-    invalid_annotation = "invalid\nannotation"
-    with pytest.raises(ValidationError):
-        _WebVTTCueVoiceSpan(annotation=invalid_annotation)
-    assert _WebVTTCueVoiceSpan(annotation=valid_annotation)
-
-    """Test that classes validation works correctly."""
-    annotation = "speaker name"
-    valid_classes = ["class1", "class2"]
-    invalid_classes = ["class\nwith\nnewlines", ""]
-    with pytest.raises(ValidationError):
-        _WebVTTCueVoiceSpan(annotation=annotation, classes=invalid_classes)
-    assert _WebVTTCueVoiceSpan(annotation=annotation, classes=valid_classes)
-
-    """Test that components validation works correctly."""
-    annotation = "speaker name"
-    valid_components = [_WebVTTCueTextSpan(text="random text")]
-    invalid_components = [123, "not a component"]
-    with pytest.raises(ValidationError):
-        _WebVTTCueVoiceSpan(annotation=annotation, components=invalid_components)
-    assert _WebVTTCueVoiceSpan(annotation=annotation, components=valid_components)
-
-    """Test valid cue voice spans."""
-    cue_span = _WebVTTCueVoiceSpan(
-        annotation="speaker",
-        classes=["loud", "clear"],
-        components=[_WebVTTCueTextSpan(text="random text")],
-    )
-
-    expected_str = "<v.loud.clear speaker>random text</v>"
-    assert str(cue_span) == expected_str
-
-    cue_span = _WebVTTCueVoiceSpan(
-        annotation="speaker",
-        components=[_WebVTTCueTextSpan(text="random text")],
-    )
-    expected_str = "<v speaker>random text</v>"
-    assert str(cue_span) == expected_str
-
-
-def test_webvtt_file():
-    """Test WebVTT files."""
-    with open("./tests/data/webvtt/webvtt_example_01.vtt", encoding="utf-8") as f:
-        content = f.read()
-        vtt = _WebVTTFile.parse(content)
-    assert len(vtt) == 13
-    block = vtt.cue_blocks[11]
-    assert str(block.timings) == "00:32.500 --> 00:33.500"
-    assert len(block.payload) == 1
-    cue_span = block.payload[0]
-    assert isinstance(cue_span, _WebVTTCueVoiceSpan)
-    assert cue_span.annotation == "Neil deGrasse Tyson"
-    assert not cue_span.classes
-    assert len(cue_span.components) == 1
-    comp = cue_span.components[0]
-    assert isinstance(comp, _WebVTTCueItalicSpan)
-    assert len(comp.components) == 1
-    comp2 = comp.components[0]
-    assert isinstance(comp2, _WebVTTCueTextSpan)
-    assert comp2.text == "Laughs"
-
-    with open("./tests/data/webvtt/webvtt_example_02.vtt", encoding="utf-8") as f:
-        content = f.read()
-        vtt = _WebVTTFile.parse(content)
-    assert len(vtt) == 4
-    reverse = (
-        "WEBVTT\n\nNOTE Copyright © 2019 World Wide Web Consortium. "
-        "https://www.w3.org/TR/webvtt1/\n\n"
-    )
-    reverse += "\n\n".join([str(block) for block in vtt.cue_blocks])
-    assert content == reverse
-
-    with open("./tests/data/webvtt/webvtt_example_03.vtt", encoding="utf-8") as f:
-        content = f.read()
-        vtt = _WebVTTFile.parse(content)
-    assert len(vtt) == 13
-    for block in vtt:
-        assert block.identifier
-    block = vtt.cue_blocks[0]
-    assert block.identifier == "62357a1d-d250-41d5-a1cf-6cc0eeceffcc/15-0"
-    assert str(block.timings) == "00:00:04.963 --> 00:00:08.571"
-    assert len(block.payload) == 1
-    assert isinstance(block.payload[0], _WebVTTCueVoiceSpan)
-    block = vtt.cue_blocks[2]
-    assert isinstance(cue_span, _WebVTTCueVoiceSpan)
-    assert block.identifier == "62357a1d-d250-41d5-a1cf-6cc0eeceffcc/16-0"
-    assert str(block.timings) == "00:00:10.683 --> 00:00:11.563"
-    assert len(block.payload) == 1
-    assert isinstance(block.payload[0], _WebVTTCueTextSpan)
-    assert block.payload[0].text == "Good."
 
 
 def test_e2e_vtt_conversions():
@@ -230,3 +40,252 @@ def test_e2e_vtt_conversions():
         )
 
         assert verify_document(doc, str(gt_path) + ".json", GENERATE)
+
+
+def _create_vtt_stream(content: str) -> DocumentStream:
+    stream = DocumentStream(name="test.vtt", stream=BytesIO(content.strip().encode()))
+    dci = _DocumentConversionInput(path_or_stream_iterator=[])
+    assert dci._guess_format(stream) == InputFormat.VTT
+
+    return stream
+
+
+def _process_vtt_doc(doc: DoclingDocument) -> str:
+    text: str = ""
+    for item in doc.texts:
+        if (
+            isinstance(item, TextItem)
+            and item.source
+            and item.source[0].kind == "track"
+        ):
+            parent = item.parent.resolve(doc)
+            if parent and isinstance(parent, GroupItem):
+                text += " "
+            text += item.text
+
+    return text.strip()
+
+
+@pytest.fixture(scope="module")
+def converter() -> DocumentConverter:
+    return DocumentConverter()
+
+
+def test_simple_two_cues_basic(converter):
+    vtt = """
+WEBVTT
+
+00:00:00.000 --> 00:00:02.000
+Hello world!
+
+00:00:02.500 --> 00:00:04.000
+Second cue.
+"""
+    stream = _create_vtt_stream(vtt)
+    doc = converter.convert(stream).document
+
+    expected = "Hello world! Second cue."
+    assert _process_vtt_doc(doc) == expected
+
+
+def test_cue_ids_present_are_ignored_in_output(converter):
+    vtt = """
+WEBVTT
+
+1
+00:00:00.000 --> 00:00:01.000
+First with ID.
+
+2
+00:00:01.250 --> 00:00:02.000
+Second with ID.
+"""
+    stream = _create_vtt_stream(vtt)
+    doc = converter.convert(stream).document
+
+    expected = "First with ID. Second with ID."
+    assert _process_vtt_doc(doc) == expected
+
+
+def test_multi_line_cue_text_preserved(converter):
+    vtt = """
+WEBVTT
+
+00:00:00.000 --> 00:00:03.000
+This is line one.
+This is line two.
+
+00:00:03.500 --> 00:00:05.000
+Another cue line one.
+Another cue line two.
+"""
+    stream = _create_vtt_stream(vtt)
+    doc = converter.convert(stream).document
+
+    expected = "This is line one. This is line two. Another cue line one. Another cue line two."
+    assert _process_vtt_doc(doc) == expected
+
+
+def test_styling_and_voice_tags_stripped(converter):
+    vtt = """
+WEBVTT
+
+00:00:00.000 --> 00:00:02.000
+<v Roger><b>Hello</b> <i>there</i><u>!</u></v>
+
+00:00:02.200 --> 00:00:04.000
+<c.red>Styled</c> and <v Ann>voiced</v> text.
+"""
+    stream = _create_vtt_stream(vtt)
+    doc = converter.convert(stream).document
+
+    # Expect tags removed but inner text retained, spacing preserved.
+    # expected = "Hello there! Styled and voiced text."
+    # TODO: temporary ground truth (issue docling-project/docling-core/#371)
+    expected = "Hello   there ! Styled  and  voiced  text."
+    assert _process_vtt_doc(doc) == expected
+
+
+def test_blank_cue_contributes_no_text(converter):
+    # First cue has text; second cue is intentionally blank (zero transcript lines).
+    vtt = """
+WEBVTT
+
+00:00:00.000 --> 00:00:02.000
+Visible text.
+
+00:00:02.500 --> 00:00:04.000
+
+"""
+    stream = _create_vtt_stream(vtt)
+    doc = converter.convert(stream).document
+
+    expected = "Visible text."
+    assert _process_vtt_doc(doc) == expected
+
+
+def test_note_blocks_are_ignored(converter):
+    vtt = """
+WEBVTT
+
+
+NOTE This is a file-level note
+It can span multiple lines.
+
+
+00:00:00.000 --> 00:00:02.000
+First cue text.
+
+
+NOTE Another note between cues
+
+
+00:00:02.500 --> 00:00:04.000
+Second cue text.
+"""
+    stream = _create_vtt_stream(vtt)
+    doc = converter.convert(stream).document
+
+    expected = "First cue text. Second cue text."
+    assert _process_vtt_doc(doc) == expected
+
+
+def test_region_block_ignored_but_region_reference_ok(converter):
+    vtt = """
+WEBVTT
+
+REGION
+id:top
+width:40%
+lines:3
+
+00:00:00.000 --> 00:00:02.000 region:top line:90% position:50% size:35% align:start
+Top region text.
+
+00:00:02.500 --> 00:00:04.000
+Normal region text.
+"""
+    stream = _create_vtt_stream(vtt)
+    doc = converter.convert(stream).document
+
+    expected = "Top region text. Normal region text."
+    assert _process_vtt_doc(doc) == expected
+
+
+def test_varied_timestamp_formats_and_settings_ignored(converter):
+    # First cue uses MM:SS.mmm; second uses HH:MM:SS.mmm and includes settings.
+    vtt = """
+WEBVTT
+
+00:01.000 --> 00:03.000
+Under one minute format.
+
+01:00:00.000 --> 01:00:02.000 line:0 position:10% align:end
+Hour format with settings.
+"""
+    stream = _create_vtt_stream(vtt)
+    doc = converter.convert(stream).document
+
+    expected = "Under one minute format. Hour format with settings."
+    assert _process_vtt_doc(doc) == expected
+
+
+def test_cue_ids_plus_multiline_with_voice_and_style(converter):
+    # Mix multiple concepts: cue IDs, multi-line text, voice tags, style tags.
+    vtt = """
+WEBVTT
+
+
+
+intro
+00:00:00.000 --> 00:00:02.000
+<v Narrator><i>Welcome</i> to the show.</v>
+<b>Enjoy</b> your time.
+
+
+
+outro
+00:00:02.500 --> 00:00:04.000
+<v Host>Goodbye</v>, see you <u>soon</u>.
+"""
+    stream = _create_vtt_stream(vtt)
+    doc = converter.convert(stream).document
+
+    # expected = "Welcome to the show. Enjoy your time. Goodbye, see you soon."
+    # TODO: temporary ground truth (issue docling-project/docling-core/#371)
+    expected = "Welcome  to the show. Enjoy  your time. Goodbye , see you  soon ."
+    assert _process_vtt_doc(doc) == expected
+
+
+def test_style_blocks_and_note_between_styles_are_ignored(converter):
+    vtt = """
+WEBVTT
+
+STYLE
+::cue {
+  background-image: linear-gradient(to bottom, dimgray, lightgray);
+  color: papayawhip;
+}
+/* Style blocks cannot use blank lines nor "dash dash greater than" */
+
+NOTE comment blocks can be used between style blocks.
+
+STYLE
+::cue(b) {
+    color: peachpuff;
+}
+
+hello
+00:00:00.000 --> 00:00:10.000
+Hello <b>world</b>.
+"""
+    stream = _create_vtt_stream(vtt)
+    with warnings.catch_warnings():
+        # STYLE and NOTE blocks should be ignored without warnings
+        warnings.simplefilter("error")
+        doc = converter.convert(stream).document
+
+    # expected = "Hello world."
+    # TODO: temporary ground truth (issue docling-project/docling-core/#371)
+    expected = "Hello  world ."
+    assert _process_vtt_doc(doc) == expected
